+++ packages/courses-backend/configs/jest-setup.ts
@@ -1,0 +1,1 @@
+import "reflect-metadata";
