+++ packages/courses-backend/doc/TODO.md
@@ -1,7 +1,9 @@
 - [x] Swagger
 - [x] Авторизация
 - [x] JWT
-- [ ] Сделать заготовку для фронта
+- [x] Сделать заготовку для фронта
+- [x] Первести на routing-controllers
+- [x] Первести на typedi
 - [ ] Сделать заготовку для деплоя
 - [ ] REST-ы по сущностям
 - [ ] Загрузка файлов
