+++ packages/courses-backend/src/__test__/authAPI.test.ts
@@ -1,5 +1,3 @@
-import "reflect-metadata";
-
 import { UserCredentials } from "@course/common";
 import { faker } from "@faker-js/faker";
 import { UserModel } from "@models/UserModel";
