+++ packages/courses-backend/src/__test__/courseAPI.test.ts
@@ -1,5 +1,3 @@
-import "reflect-metadata";
-
 import { Course, UserCredentials } from "@course/common";
 import { faker } from "@faker-js/faker";
 import { CourseModel } from "@models/CourseModel";
